requirements.txt
Flask
google-generativeai
gunicorn
python-dotenv
requests
